new structure

+++ Generic Sorting Algorithms/shell/shellsort.h
@@ -1,11 +1,13 @@
 #ifndef __SHELLSORT_H
 #define __SHELLSORT_H
 
+#include <stddef.h>
+
 typedef int (*Comparator)(const void *, const void *);
 
 typedef int (*step_generator)(int);
 
-void shellsort(void *arr, int array_size, int element_size, Comparator cmp,
+void shellsort(void *arr, size_t array_size, size_t element_size, Comparator cmp,
                int gaps[], int number_of_gaps);
 
 void generate_gaps_fib(int gaps[], int *number_of_gaps, int array_size);
